add reboot tool
This is a tool for easy upload in Arduino environment. The tool can disconnect power and hard reset the target, and pull D+ to enter bootloader. Quite useful for development with USB device, or the chip such as CH549 that can not jump into bootloader. Also the target can pull down the P1.7 pin for peripheral switch.